+++ lib/paper_3d.dart
@@ -4,7 +4,7 @@
 library paper_3d;
 
 export 'src/paper_world/paper_world_widget.dart';
-export 'src/util/blocToSubject.dart';
+export 'src/util/bloc_to_subject.dart';
 export 'src/util/screen_size.dart';
 export 'src/world_asset/world_asset.dart';
 export 'src/world_asset/world_asset_internal/world_asset_render_model.dart';
@@ -14,4 +14,4 @@ export 'src/world_asset/animation/default_asset_animation.dart';
 export 'src/world_asset/animation/world_asset_model.dart';
 export 'src/base_camera_model.dart';
 export 'src/scene.dart';
-export 'src/svg-shape-border.dart';
+export 'src/svg_shape_border.dart';