+++ lib/src/scene.dart
@@ -8,16 +8,16 @@ class Scene {
   final Size screenSize;
   final Size size;
 
-  late final Matrix4 toOpenGlCoords;
-  late final Matrix4 toFlutterCoords;
+  late final Matrix4 toOpenGlCoordinates;
+  late final Matrix4 toFlutterCoordinates;
 
   Scene(this.screenSize, this.size) {
-    toOpenGlCoords = Matrix4.compose(Vector3(-1.0, 1.0, 0.0), Quaternion.axisAngle(Vector3(1.0, 0.0, 0.0), pi), Vector3(2 / size.width, 2 / size.height, 1.0));
-    toFlutterCoords = Matrix4.tryInvert(toOpenGlCoords)!;
+    toOpenGlCoordinates = Matrix4.compose(Vector3(-1.0, 1.0, 0.0), Quaternion.axisAngle(Vector3(1.0, 0.0, 0.0), pi), Vector3(2 / size.width, 2 / size.height, 1.0));
+    toFlutterCoordinates = Matrix4.tryInvert(toOpenGlCoordinates)!;
   }
 
   widthToGL(double width) => width * 2 / size.width;
-  heightToGL(double width) => width * 2 / size.height;
+  heightToGL(double height) => height * 2 / size.height;
 
   modelToSceneCenterTranslation(Size modelSize) => Matrix4.translation(Vector3(
     widthToGL((size.width - modelSize.width)) / 2,